+++ src/content.ts
@@ -1,3 +1,9 @@
+export const homeText =
+  'Et fond er en kollektiv invistering der flere går sammen å plasserer penger i verdipapirmarkedet. Dette kalles ofte for kollektiv sparing, og man kan se for seg at et fond er en kurv med investeringer. Online har et slikt fond, dette betyr at ulike parter innvesterer i Online, noe vi setter stor pris på. Under finner du en graf, denne grafen er en oversikt over utviklingen av Online sitt fond. Grafen inkluderer også en graf for OSEBX, som er hovedindeksen på Oslo Børs. Disse grafene kan sammenlignes for å se hvordan Online sitt fond har utviklet seg i forhold til hovedindeksen på Oslo Børs. Under finner du en graf, denne grafen er en oversikt over den relative utviklingen til Online sitt fond.';
+
+export const applicationText =
+  'Alle Onlines medlemmer kan søke Onlines fond om penger. Søknaden skal være velbegrunnet og ha som hensikt å komme flest mulig medlemmer av Online til gode. For at Fondstyre skal kunne ta en god avgjørelse trenger vi en helhetlig forståelse av hvem dere er og hva dere gjør. Skriv gjerne litt for mye, enn litt for lite. Søknaden er kun gyldig dersom det søkes om mellom 10 000 kr og 100 000 kr.';
+
 export const aboutUsText =
   'Onlinefondet er en engasjert og livlig studentorganisasjon ved NTNU, dedikert til å fremme studentenes interesser og berike deres akademiske og sosiale opplevelser. Gjennom vårt fond, ledet av et dedikert fondstyre, gir vi økonomisk støtte til initiativer som sikrer at medlemmenes behov og interesser blir ivaretatt. Fra å finansiere nytt musikkutstyr for vårt linjeforeningsband, til å støtte surfeturer og sosiale arrangementer, streber vi etter å gi tilbake til vår studentkommunitet. Med en rik historie og et engasjert medlemskap, er vi stolte av vår evne til å kombinere det faglige med det sosiale, og skape varige minner for alle Onlinere.';
